Refactor register mutation
