--- src/modules/auth/SignUpForm/signUp.gql
@@ -1,11 +1,0 @@
-mutation Register($payload: RegisterDetailInput!) {
-  register(detail: $payload) {
-    accessToken
-    user {
-      id
-      status
-      username
-      email
-    }
-  }
-}
